refactor(announce): extract build/sign/verify into shared internal/announce package
