+++ cmd/sam-node/announce.go
@@ -19,62 +19,10 @@ import (
 	"time"
 
 	"github.com/google/sam/api"
+	"github.com/google/sam/internal/announce"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
-	"github.com/libp2p/go-libp2p/core/crypto"
-	"github.com/libp2p/go-libp2p/core/peer"
 	"google.golang.org/protobuf/proto"
 )
-
-// serviceAnnounceTTL is the lifetime a catalog should keep an entry before
-// eviction; ~3x the 5-min reprovide tick so one dropped announce is tolerated.
-const serviceAnnounceTTL = 15 * time.Minute
-
-// buildServiceAnnounce assembles an unsigned announce for a local service.
-func buildServiceAnnounce(info *api.ServiceInfo, peerID peer.ID, addrs []string, now time.Time, ttl time.Duration) *api.ServiceAnnounce {
-	return &api.ServiceAnnounce{
-		Type:      info.Type,
-		Name:      info.Name,
-		PeerId:    peerID.String(),
-		Addrs:     addrs,
-		Timestamp: now.UnixMilli(),
-		TtlMs:     ttl.Milliseconds(),
-	}
-}
-
-// signServiceAnnounce signs the announce over its signature-cleared marshalling.
-func signServiceAnnounce(priv crypto.PrivKey, a *api.ServiceAnnounce) error {
-	a.Signature = nil
-	data, err := proto.Marshal(a)
-	if err != nil {
-		return err
-	}
-	sig, err := priv.Sign(data)
-	if err != nil {
-		return err
-	}
-	a.Signature = sig
-	return nil
-}
-
-// verifyServiceAnnounce checks the signature against the key derived from PeerId.
-func verifyServiceAnnounce(a *api.ServiceAnnounce) (bool, error) {
-	pid, err := peer.Decode(a.PeerId)
-	if err != nil {
-		return false, err
-	}
-	pub, err := pid.ExtractPublicKey()
-	if err != nil {
-		return false, err
-	}
-	sig := a.Signature
-	a.Signature = nil
-	data, err := proto.Marshal(a)
-	a.Signature = sig // restore
-	if err != nil {
-		return false, err
-	}
-	return pub.Verify(data, sig)
-}
 
 // joinTopic returns the cached topic, joining and caching it on first use.
 func (n *SamNode) joinTopic(name string) (*pubsub.Topic, error) {
@@ -115,8 +63,8 @@ func (n *SamNode) announceServices(ctx context.Context) {
 
 	now := time.Now()
 	for _, info := range n.services.List(api.ServiceType_SERVICE_TYPE_UNSPECIFIED) {
-		ann := buildServiceAnnounce(info, n.Host.ID(), addrs, now, serviceAnnounceTTL)
-		if err := signServiceAnnounce(priv, ann); err != nil {
+		ann := announce.Build(info, n.Host.ID(), addrs, now, announce.TTL)
+		if err := announce.Sign(priv, ann); err != nil {
 			logger.Warnf("[Announce] sign %s: %v", info.Name, err)
 			continue
 		}
